NEW: Improve portability
The server launcher is now a composer bin command, rather than a
SilverStripe build tasks, which makes it compatible with both 3.x
and 4.x without changes.

Class loading utilises composer, and the serve module can be either
at <root>/server or vendor/silverstripe/serve, which is more orthodox.

The type: silverstripe-module marker is no longer needed as this just
controls the install location of the module.

Reliance on BASE_PATH and FRAMEWORK_PARH, which are set by the composer
autoloader whenever silverstripe/framework is included, make this more
resilient to filesystem structure changes.

The up-shot of all this is that serve can be listed as a dev-dependency
of a SilverStripe module and used to do integration testing of that
module, provided other work around SilverStripe’s filesystem decoupling
also happens.