NEW: Allow path argument to the factory constructor.
Normally, BASE_PATH is used as the path because this is designed for
SilverStripe installs. However, sometimes in cases such as tests, this
needs to be set.

I’ve opted to put it in the factory constructor rather than the
launchServer() options, because it’s more a property of the operating
environment than the server requested.